Cache path lookups when running commands
Until now, rc used to look for an executable in each member
of $path for each invocation. That's a lot of syscalls!

This new cache prevents that. In order to stay an minimalist
as possible, I chose not to add a "hash" builtin to reset the
cache, like in most Bourne shells. Instead, the cache will
automatically invalidate entries where the command is not
executable anymore. The cache also resets itself when $path
is modified. This should avoid any bad surprise 🤞